Added Working Commodity Controller and all related classes and html pages.

+++ src/main/java/lv/pakit/repo/ICommodityRepo.java
@@ -3,5 +3,5 @@ package lv.pakit.repo;
 import lv.pakit.model.Commodity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICommodityRepo extends JpaRepository<Commodity, Integer> {
+public interface ICommodityRepo extends JpaRepository<Commodity, Long> {
 }
